week4 answers
